Add friendly backup names to barman
Adds the `--name` option to `barman backup` which can be used to store
a user-assigned name in the backup metadata. This name can then be used
in barman commands in place of the backup ID in order to perform a given
operation on that backup.

This is implemented by storing the backup name as a new field in the
`backup.info` file named `backup_name`. The `backup_name` field is not
written to the `backup.info` file or included in the `InfoFile` output
if it is `None` - this means that `backup.info` files will only break
compatibility with older versions of Barman if they opt-in to the
feature by using the `--name` option.

The `barman list-backups` and `barman show-backup` commands are updated
so that their output includes the backup name (if set) in both the plain
and JSON output formats.

The following limitations apply to backup names:

  - Backup names which have the same format as the backup ID (ISO 8601
    YYYYMMDDTHHMMSS) are not allowed.
  - Backup names cannot use the backup ID shortcuts (currently: latest,
    last, earliest, oldest and last-failed).

If multiple backups have the same backup name this will be detected when
resolving the name and result in an error. The backup IDs of the
matching backups are provided so users can resolve the issue using the
backup IDs instead.

Part 1 of 2 closing #690.