Add friendly backup names to barman cloud
Adds the `--name` option to `barman-cloud-backup` which can be used to
store a user-assigned name in the backup metadata. As with regular
barman commands, this backup name can then be used in place of the
backup ID in order to perform a given operation on that backup.

The feature works the same way as with regular barman, using the
`backup.info` field introduced in the previous commit. The difference
in implementation is that the `backup.info` files are in cloud storage
so resolving the name to a backup ID must happen in the
`CloudBackupCatalog`.

As with the feature in regular barman, names cannot be backup IDs or
backup ID shortcuts and if multiple backups match a backup ID the user
will encounter an error along with the list of matching backup IDs.

Part 2 of 2 closing #690.